Private Documentation

Legendre.Scalar — Type

mutable struct Scalar{T} <: AbstractArray{T,0}

A 0-dimensional (scalar) object, much like a RefValue{T}, but one which is an abstract array and assignable under broadcasting operations.

Legendre.Plm_00 — Function

Plm_00(::N, ::Type{T}) where {N<:AbstractLegendreNorm, T}

Returns the initial condition $P_0^0(x)$ for the associated Legendre recursions based on the normalization choice N for numeric type T.

Legendre.Plm_α — Function

Plm_α(norm::N, ::Type{T}, l::Integer, m::Integer) where {N<:AbstractLegendreNorm, T}

Returns the coefficient $α_ℓ^m$ for the two-term recursion relation

\[ P_{ℓ+1}^{m}(x) = α_{ℓ+1}^m x P_ℓ^m(x) - β_{ℓ+1}^m P_{ℓ-1}^m(x)\]

where $α_ℓ^m$ is appropriate for the choice of normalization N.

Legendre.Plm_β — Function

Plm_β(norm::N, ::Type{T}, l::Integer, m::Integer) where {N<:AbstractLegendreNorm, T}

Returns the coefficient $β_ℓ^m$ for the two-term recursion relation

\[ P_{ℓ+1}^{m}(x) = α_{ℓ+1}^m x P_ℓ^m(x) - β_{ℓ+1}^m P_{ℓ-1}^m(x)\]

where $β_ℓ^m$ is appropriate for the choice of normalization N.

Legendre.Plm_μ — Function

Plm_μ(norm::N, ::Type{T}, l::Integer) where {N<:AbstractLegendreNorm, T}

Returns the coefficient $μ_ℓ$ for the single-term recursion relation

\[ P_{ℓ+1}^{ℓ+1}(x) = -μ_{ℓ+1} \sqrt{1-x^2} P_ℓ^ℓ(x)\]

where $μ_ℓ$ is appropriate for the choice of normalization N.

Legendre.Plm_ν — Function

Plm_ν(norm::N, ::Type{T}, l::Integer) where {N<:AbstractLegendreNorm, T}

Returns the coefficient $ν_ℓ$ for the single-term recursion relation

\[ P_{ℓ+1}^ℓ(x) = ν_{ℓ+1} x P_ℓ^ℓ(x)\]

where $ν_ℓ$ is appropriate for the choice of normalization N.

Legendre.boundscheck_hook — Function

boundscheck_hook(norm::AbstractLegendreNorm, lmax, mmax)

A bounds-checking hook executed at the beginning of each legendre! call to permit a normalization norm to validate that the given maximum $(ℓ,m)$ will be within the ability to satisfy. The default case always returns nothing. A custom normalization should throw an error if lmax or mmax is out of bounds or return nothing otherwise.

For example, the precomputed coefficients of LegendreNormCoeff are limited to a given domain at time of construction and cannot be used to calculate terms to arbitrary orders/degrees.