API Reference

Functions

Legendre.legendre — Function

p = legendre(norm::AbstractLegendreNorm, l::DimOrInd, m::DimOrInd, x)

Computes the associated Legendre polynomials $N_ℓ^m P_ℓ^m(x)$ of degree(s) l and order(s) m at x for the normalization scheme norm.

If l and m are integers, returns a single value.

If l is a range 0:lmax and m an integer, returns the vector of values for order m and all degrees 0 ≤ l ≤ lmax.

If l is a range 0:lmax and m is a range 0:mmax, returns the matrix of values for all degrees 0 ≤ l ≤ lmax and orders 0 ≤ m ≤ mmax.

Note that in both the second and third cases, the ranges must have a first index of 0.

Legendre.legendre! — Function

legendre!(norm::AbstractLegendreNorm, Λ, l::Integer, m::Integer, x)

Fills the array Λ with the Legendre polynomial values $N_ℓ^m P_ℓ^m(x)$ up to/of degree(s) l and order(s) m for the normalization scheme norm. Λ must be an array with between 0 and 2 more dimensions than x, with the leading dimensions having the same shape as x.

If ndims(Λ) == ndims(x), then Λ is filled with the polynomial values at x for degree l and order m.

If ndims(Λ) == ndims(x) + 1, then l is interpreted as lmax, and Λ filled with polynomial values for all degrees 0 ≤ l ≤ lmax of order m.

If ndims(Λ) == ndims(x) + 2, then l is interpreted as lmax and m as mmax, and Λ is filled with polynomial values for all degrees 0 ≤ l ≤ lmax and orders 0 ≤ m ≤ min(mmax, l).

Legendre.Nlm — Function

N = Nlm([T=Float64], l, m)

Computes the normalization constant

\[ N_ℓ^m ≡ \sqrt{\frac{2ℓ+1}{4π} \frac{(ℓ-m)!}{(ℓ+m)!}}\]

which defines the Spherical Harmonic normalized functions $λ_ℓ^m(x)$ in terms of the standard unit normalized $P_ℓ^m(x)$

\[ λ_ℓ^m(x) ≡ N_ℓ^m P_ℓ^m(x)\]

using numbers of type T.

See also Plm and λlm.

Normalizations

Legendre.AbstractLegendreNorm — Type

abstract type AbstractLegendreNorm end

Abstract trait supertype for normalization conditions of the Associated Legendre polynomials.

Legendre.LegendreUnitNorm — Type

struct LegendreUnitNorm <: AbstractLegendreNorm end

Trait type denoting the unit normalization of the associated Legendre polynomials.

Legendre.LegendreSphereNorm — Type

struct LegendreSphereNorm <: AbstractLegendreNorm end

Trait type denoting the spherical-harmonic normalization of the associated Legendre polynomials.

Legendre.LegendreNormCoeff — Type

struct LegendreNormCoeff{N<:AbstractLegendreNorm,T<:Real} <: AbstractLegendreNorm

Precomputed recursion relation coefficients for the normalization N and value type T.

Example

julia> LegendreNormCoeff{LegendreSphereNorm,Float64}(1)
LegendreNormCoeff{LegendreSphereNorm,Float64} for lmax = 1, mmax = 1 with coefficients:
    μ: [0.0, 1.22474]
    ν: [1.73205, 2.23607]
    α: [0.0 0.0; 1.73205 0.0]
    β: [0.0 0.0; -0.0 0.0]

Aliases

The following functors are constant aliases to the underlying normalization types which have been made callable via an appropriate call overload.

Legendre.Plm — Constant

p = Plm(l, m, x)

Computes the associated Legendre polynomials using unit normalization; equivalent to p = legendre(LegendreUnitNorm(), l, m, x).

Legendre.λlm — Constant

λ = λlm(l, m, x)

Computes the associated Legendre polynomials using spherical-harmonic normalization; equivalent to λ = legendre(LegendreSphereNorm(), l, m, x).

Legendre.Plm! — Constant

Plm!(P, l, m, x)

Fills the array P with the unit-normalized associated Legendre polynomial values $P_ℓ^m(x)$; equivalent to legendre!(LegendreUnitNorm(), P, l, m, x).

Legendre.λlm! — Constant

λlm!(Λ, l, m, x)

Fills the array Λ with the spherical-harmonic normalized associated Legendre polynomial values $λ_ℓ^m(x)$; equivalent to legendre!(LegendreSphereNorm(), P, l, m, x).

There are also aliases for pre-computed coefficients of the provided normalizations.

Legendre.LegendreUnitCoeff — Type

LegendreUnitCoeff{T}

Precomputed recursion relation coefficients for the standard unit normalization. Alias for LegendreNormCoeff{LegendreUnitNorm,T}.

Legendre.LegendreSphereCoeff — Type

LegendreSphereCoeff{T}

Table type of precomputed recursion relation coefficients for the spherical harmonic normalization. Alias for LegendreNormCoeff{LegendreSphereNorm,T}.

Normalization Interface

The following functions are unexported but considered a public API for interacting with and defining additional normalizations.

Legendre.initcond — Function

initcond(::N, ::Type{T}) where {N<:AbstractLegendreNorm, T}

Returns the initial condition $P_0^0(x)$ for the associated Legendre recursions based on the normalization choice N for numeric type T.

Legendre.coeff_μ — Function

coeff_μ(norm::N, ::Type{T}, l::Integer) where {N<:AbstractLegendreNorm, T}

Returns the coefficient $μ_ℓ$ for the single-term recursion relation

\[ P_ℓ^ℓ(x) = -μ_ℓ \sqrt{1-x^2} P_{ℓ-1}^{ℓ-1}(x)\]

where $μ_ℓ$ is appropriate for the choice of normalization N.

Legendre.coeff_ν — Function

coeff_ν(norm::N, ::Type{T}, l::Integer) where {N<:AbstractLegendreNorm, T}

Returns the coefficient $ν_ℓ$ for the single-term recursion relation

\[ P_ℓ^{ℓ-1}(x) = ν_ℓ x P_{ℓ-1}^{ℓ-1}(x)\]

where $ν_ℓ$ is appropriate for the choice of normalization N.

Legendre.coeff_α — Function

coeff_α(norm::N, ::Type{T}, l::Integer, m::Integer) where {N<:AbstractLegendreNorm, T}

Returns the coefficient $α_ℓ^m$ for the two-term recursion relation

\[ P_ℓ^m(x) = α_ℓ^m x P_{ℓ-1}^m(x) - β_ℓ^m P_{ℓ-2}^m(x)\]

where $α_ℓ^m$ is appropriate for the choice of normalization N.

Legendre.coeff_β — Function

coeff_β(norm::N, ::Type{T}, l::Integer, m::Integer) where {N<:AbstractLegendreNorm, T}

Returns the coefficient $β_ℓ^m$ for the two-term recursion relation

\[ P_ℓ^m(x) = α_ℓ^m x P_{ℓ-1}^m(x) - β_ℓ^m P_{ℓ-2}^m(x)\]

where $β_ℓ^m$ is appropriate for the choice of normalization N.

Legendre.boundscheck_hook — Function

boundscheck_hook(norm::AbstractLegendreNorm, lmax, mmax)

A bounds-checking hook executed at the beginning of each legendre! call to permit a normalization norm to validate that the given maximum $(ℓ,m)$ will be within the ability to satisfy. The default case always returns nothing. A custom normalization should throw an error if lmax or mmax is out of bounds or return nothing otherwise.

For example, the precomputed coefficients of LegendreNormCoeff are limited to a given domain at time of construction and cannot be used to calculate terms to arbitrary orders/degrees.